adds functionality to remove button
