+++ src/main/kotlin/eu/florian_fuhrmann/musictimedtriggers/gui/views/app/browser/TriggerTemplatesList.kt
@@ -21,10 +21,7 @@ import eu.florian_fuhrmann.musictimedtriggers.gui.dialogs.alerts.AlertCreator
 import eu.florian_fuhrmann.musictimedtriggers.gui.uistate.MainUiState
 import eu.florian_fuhrmann.musictimedtriggers.gui.uistate.browser.BrowserState
 import eu.florian_fuhrmann.musictimedtriggers.gui.uistate.browser.BrowserTemplate
-import eu.florian_fuhrmann.musictimedtriggers.gui.views.app.editor.timeline.managers.TriggerSelectionManager
-import eu.florian_fuhrmann.musictimedtriggers.gui.views.app.editor.timeline.redrawTimeline
 import eu.florian_fuhrmann.musictimedtriggers.project.ProjectManager
-import eu.florian_fuhrmann.musictimedtriggers.triggers.sequence.TriggerSequenceLine
 import eu.florian_fuhrmann.musictimedtriggers.utils.icons.MttIcons
 import eu.florian_fuhrmann.musictimedtriggers.utils.color.getContrasting
 import org.jetbrains.jewel.foundation.modifier.onHover
@@ -70,7 +67,7 @@ fun TriggerTemplatesList() {
                                 ""
                             },
                     ) {
-                        removeSelectedTemplates(browserState)
+                        browserState.removeSelectedTemplates()
                     },
                     ContextMenuItem("Copy") {
                         browserState.copy()
@@ -98,7 +95,7 @@ fun TriggerTemplatesList() {
                         if (it.type != KeyEventType.KeyUp) return@onKeyEvent false
                         if (it.key == Key.Backspace || it.key == Key.Delete) {
                             // remove all selected triggers
-                            removeSelectedTemplates(browserState, it.isShiftPressed && it.isAltPressed)
+                            browserState.removeSelectedTemplates(it.isShiftPressed && it.isAltPressed)
                             return@onKeyEvent true
                         }
                         return@onKeyEvent false
@@ -282,41 +279,6 @@ fun TriggerTemplateItem(
                 )
             }
         }
-    }
-}
-
-private fun removeSelectedTemplates(browserState: BrowserState, skipConfirmation: Boolean = false) {
-    // ensure that at least one template is selected
-    if (browserState.selectedTemplates.isEmpty()) {
-        return
-    }
-    // get current project
-    val project = ProjectManager.currentProject ?: throw IllegalStateException("No project currently open")
-    // collect triggers to remove
-    val selectedTemplates = browserState.selectedTemplates.map { it.getTriggerTemplate() }
-    // search for usages of the selected templates
-    val usages = project.triggersManager.searchUsagesOfTriggerTemplates(project, selectedTemplates)
-    // create onConfirm function
-    val onConfirm: () -> Unit = {
-        // remove placed triggers in usages and collect set of affected lines
-        val affectedLines = mutableSetOf<TriggerSequenceLine>()
-        usages.forEach {
-            TriggerSelectionManager.deselectTrigger(it.placedTrigger, false)
-            it.line.removeTrigger(it.placedTrigger)
-            affectedLines.add(it.line)
-        }
-        // redraw timeline because some placed triggers currently visible might have been removed
-        redrawTimeline()
-        // save the affected lines
-        affectedLines.forEach { it.saveToFile() }
-        // remove the templates
-        project.triggersManager.removeTriggerTemplates(selectedTemplates) // also saves the affected groups
-    }
-    // show confirmation dialog if needed
-    if (skipConfirmation) {
-        onConfirm.invoke()
-    } else {
-        DialogManager.alert(AlertCreator.createUsagesAlert(true, selectedTemplates, usages, onConfirm))
     }
 }
